// src/yeying/api/bulletin/bulletin.ts
export interface BulletinListRequest {
  header: MessageHeader | undefined;
  body: BulletinListRequestBody | undefined;
}

export interface BulletinListRequestBody {
  code: BulletinCodeEnum;
  language: LanguageCodeEnum;
  page: RequestPage | undefined;
}

export interface BulletinListResponse {
  header: MessageHeader | undefined;
  body: BulletinListResponseBody | undefined;
}

export interface BulletinListResponseBody {
  status: ResponseStatus | undefined;
  solutions: SolutionMetadata[];
  page: ResponsePage | undefined;
}
// ...
export const BulletinListRequest: MessageFns<BulletinListRequest> = {
  encode(message: BulletinListRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      BulletinListRequestBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BulletinListRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBulletinListRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = BulletinListRequestBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BulletinListRequest {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? BulletinListRequestBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: BulletinListRequest): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = BulletinListRequestBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BulletinListRequest>, I>>(base?: I): BulletinListRequest {
    return BulletinListRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BulletinListRequest>, I>>(object: I): BulletinListRequest {
    const message = createBaseBulletinListRequest();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? BulletinListRequestBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseBulletinListRequestBody(): BulletinListRequestBody {
  return { code: 0, language: 0, page: undefined };
}

export const BulletinListRequestBody: MessageFns<BulletinListRequestBody> = {
  encode(message: BulletinListRequestBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.code !== 0) {
      writer.uint32(8).int32(message.code);
    }
    if (message.language !== 0) {
      writer.uint32(16).int32(message.language);
    }
    if (message.page !== undefined) {
      RequestPage.encode(message.page, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BulletinListRequestBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBulletinListRequestBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.code = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.language = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.page = RequestPage.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BulletinListRequestBody {
    return {
      code: isSet(object.code) ? bulletinCodeEnumFromJSON(object.code) : 0,
      language: isSet(object.language) ? languageCodeEnumFromJSON(object.language) : 0,
      page: isSet(object.page) ? RequestPage.fromJSON(object.page) : undefined,
    };
  },

  toJSON(message: BulletinListRequestBody): unknown {
    const obj: any = {};
    if (message.code !== 0) {
      obj.code = bulletinCodeEnumToJSON(message.code);
    }
    if (message.language !== 0) {
      obj.language = languageCodeEnumToJSON(message.language);
    }
    if (message.page !== undefined) {
      obj.page = RequestPage.toJSON(message.page);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BulletinListRequestBody>, I>>(base?: I): BulletinListRequestBody {
    return BulletinListRequestBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BulletinListRequestBody>, I>>(object: I): BulletinListRequestBody {
    const message = createBaseBulletinListRequestBody();
    message.code = object.code ?? 0;
    message.language = object.language ?? 0;
    message.page = (object.page !== undefined && object.page !== null)
      ? RequestPage.fromPartial(object.page)
      : undefined;
    return message;
  },
};

function createBaseBulletinListResponse(): BulletinListResponse {
  return { header: undefined, body: undefined };
}

export const BulletinListResponse: MessageFns<BulletinListResponse> = {
  encode(message: BulletinListResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      BulletinListResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BulletinListResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBulletinListResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = BulletinListResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BulletinListResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? BulletinListResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: BulletinListResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = BulletinListResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BulletinListResponse>, I>>(base?: I): BulletinListResponse {
    return BulletinListResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BulletinListResponse>, I>>(object: I): BulletinListResponse {
    const message = createBaseBulletinListResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? BulletinListResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseBulletinListResponseBody(): BulletinListResponseBody {
  return { status: undefined, solutions: [], page: undefined };
}

export const BulletinListResponseBody: MessageFns<BulletinListResponseBody> = {
  encode(message: BulletinListResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    for (const v of message.solutions) {
      SolutionMetadata.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.page !== undefined) {
      ResponsePage.encode(message.page, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BulletinListResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBulletinListResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.solutions.push(SolutionMetadata.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.page = ResponsePage.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BulletinListResponseBody {
    return {
      status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined,
      solutions: globalThis.Array.isArray(object?.solutions)
        ? object.solutions.map((e: any) => SolutionMetadata.fromJSON(e))
        : [],
      page: isSet(object.page) ? ResponsePage.fromJSON(object.page) : undefined,
    };
  },

  toJSON(message: BulletinListResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    if (message.solutions?.length) {
      obj.solutions = message.solutions.map((e) => SolutionMetadata.toJSON(e));
    }
    if (message.page !== undefined) {
      obj.page = ResponsePage.toJSON(message.page);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BulletinListResponseBody>, I>>(base?: I): BulletinListResponseBody {
    return BulletinListResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BulletinListResponseBody>, I>>(object: I): BulletinListResponseBody {
    const message = createBaseBulletinListResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    message.solutions = object.solutions?.map((e) => SolutionMetadata.fromPartial(e)) || [];
    message.page = (object.page !== undefined && object.page !== null)
      ? ResponsePage.fromPartial(object.page)
      : undefined;
    return message;
  },
};
// ...
export interface BulletinServer extends UntypedServiceImplementation {
  list: handleUnaryCall<BulletinListRequest, BulletinListResponse>;
}
// ...
type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };